62-壁を描画できるように (#63)
* update(renderring): レンダリングの関数を作成

* update(render): 仮で壁が描画できるものを追加

しかし、調査を行い実装を確認する必要がある。

* fix(render): 壁の描画は出来ているが、魚眼効果で屈曲している

* fix: angleをすべてdoubleに変更

* fix(render): 描画角度を修正

* fix(render): 魚眼を補正

+++ cmd/cub3D/movement/trig_degree.c
@@ -6,13 +6,13 @@
 /*   By: saraki <saraki@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 04:18:25 by syamasaw          #+#    #+#             */
-/*   Updated: 2024/10/26 22:44:20 by saraki           ###   ########.fr       */
+/*   Updated: 2024/10/28 14:51:59 by saraki           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-double	cos_degree(int angle)
+double	cos_degree(double angle)
 {
 	double	rad;
 	double	ret;
@@ -24,7 +24,7 @@ double	cos_degree(int angle)
 	return (ret);
 }
 
-double	sin_degree(int angle)
+double	sin_degree(double angle)
 {
 	double	rad;
 	double	ret;
