+++ cmd/cub3D/movement/update_coodinate.c
@@ -6,27 +6,27 @@
 /*   By: saraki <saraki@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/18 07:43:41 by syamasaw          #+#    #+#             */
-/*   Updated: 2024/10/26 23:51:11 by saraki           ###   ########.fr       */
+/*   Updated: 2024/10/28 16:56:12 by saraki           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-static void	set_player_to_edge(t_dda *dda_ret, t_data *data, int mv_angle);
+static void	set_player_to_edge(t_dda *dda_ret, t_data *data, double mv_angle);
 
 void	update_coordinate(t_data *data, int key)
 {
-	int			angle;
+	double		angle;
 	t_dda		dda_ret;
 	t_player	adjust_player;
 
 	angle = data->player.angle;
 	if (key == KEY_D)
-		angle += 90;
+		angle += 90.0;
 	else if (key == KEY_A)
-		angle -= 90;
+		angle -= 90.0;
 	else if (key == KEY_S)
-		angle += 180;
+		angle += 180.0;
 	angle = regulate_angle(angle);
 	adjust_player.pos_x = data->player.pos_x;
 	adjust_player.pos_y = data->player.pos_y;
@@ -40,7 +40,7 @@ void	update_coordinate(t_data *data, int key)
 	set_player_to_edge(&dda_ret, data, angle);
 }
 
-static void	set_player_to_edge(t_dda *dda, t_data *d, int mv_angle)
+static void	set_player_to_edge(t_dda *dda, t_data *d, double mv_angle)
 {
 	double		next_distance;
 	double		next[2];
